[refactor] List Item style 공통화
