+++ src/components/List.style.jsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const List = styled.ul({
+  display: 'flex',
+  margin: '1em 0',
+  padding: 0,
+  listStyle: 'none',
+});
+
+export default List;
